+++ core/ceremony/lottery_test.go
@@ -3,6 +3,7 @@ package ceremony
 import (
 	"encoding/binary"
 	"github.com/google/tink/go/subtle/random"
+	"github.com/idena-network/idena-go/common"
 	"github.com/stretchr/testify/require"
 	"math/rand"
 	"testing"
@@ -12,16 +13,16 @@ func Test_GetFirstAuthorsDistribution(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 100)
 
-	candidates, _, _ := makeCandidatesWithFlips(seed, 120, 100, 3)
-	authors := getAuthorsIndexes(candidates)
-	authorsPerCandidate, _ := getFirstAuthorsDistribution(authors, candidates, seed, 7)
+	candidates := makeCandidatesWithFlips(seed, 120, 100, 3)
+	authors := getAuthorsIndexes(candidates[0].candidates)
+	authorsPerCandidate, _ := getFirstAuthorsDistribution(authors, candidates[0].candidates, seed, 7)
 	for _, item := range authorsPerCandidate {
 		require.Equal(t, 7, len(item))
 	}
 
-	candidates, _, _ = makeCandidatesWithFlips(seed, 10000, 5000, 3)
-	authors = getAuthorsIndexes(candidates)
-	authorsPerCandidate, _ = getFirstAuthorsDistribution(authors, candidates, seed, 7)
+	candidates = makeCandidatesWithFlips(seed, 10000, 5000, 3)
+	authors = getAuthorsIndexes(candidates[0].candidates)
+	authorsPerCandidate, _ = getFirstAuthorsDistribution(authors, candidates[0].candidates, seed, 7)
 
 	for _, item := range authorsPerCandidate {
 		require.Equal(t, 7, len(item))
@@ -32,10 +33,10 @@ func Test_Case1(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 100)
 
-	candidates, flipsPerAuthor, flips := makeCandidatesWithFlips(seed, 150, 100, 3)
+	candidates := makeCandidatesWithFlips(seed, 150, 100, 3)
 
-	authorsPerCandidate, _ := GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate := GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate, candidates[0].flipsPerAuthor, candidates[0].flips, seed, 7)
 
 	for _, item := range shortFlipsPerCandidate {
 		require.Equal(t, 7, len(item))
@@ -50,10 +51,11 @@ func Test_Case2(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 500)
 
-	candidates, flipsPerAuthor, flips := makeCandidatesWithFlips(seed, 300, 100, 3)
+	candidates := makeCandidatesWithFlips(seed, 300, 100, 3)
 
-	authorsPerCandidate, _ := GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate := GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate,
+		candidates[0].flipsPerAuthor, candidates[0].flips, seed, 7)
 
 	for _, item := range shortFlipsPerCandidate {
 		require.Equal(t, 7, len(item))
@@ -68,18 +70,20 @@ func Test_Case3(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 500)
 
-	candidates, flipsPerAuthor, flips := makeCandidatesWithFlips(seed, 5, 0, 3)
+	candidates := makeCandidatesWithFlips(seed, 5, 0, 3)
 
-	authorsPerCandidate, _ := GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate := GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate, candidates[0].flipsPerAuthor,
+		candidates[0].flips, seed, 7)
 
 	require.Equal(t, 5, len(shortFlipsPerCandidate))
 	require.Equal(t, 5, len(longFlipsPerCandidate))
 
-	candidates, flipsPerAuthor, flips = makeCandidatesWithFlips(seed, 0, 0, 3)
+	candidates = makeCandidatesWithFlips(seed, 0, 0, 3)
 
-	authorsPerCandidate, _ = GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate = GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate = GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate = GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate, candidates[0].flipsPerAuthor,
+		candidates[0].flips, seed, 7)
 	require.Equal(t, 0, len(shortFlipsPerCandidate))
 	require.Equal(t, 0, len(longFlipsPerCandidate))
 }
@@ -88,10 +92,11 @@ func Test_Case4(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 500)
 
-	candidates, flipsPerAuthor, flips := makeCandidatesWithFlips(seed, 12, 1, 1)
+	candidates := makeCandidatesWithFlips(seed, 12, 1, 1)
 
-	authorsPerCandidate, _ := GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate := GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate,
+		candidates[0].flipsPerAuthor, candidates[0].flips, seed, 7)
 
 	for _, item := range shortFlipsPerCandidate {
 		require.Equal(t, 1, len(item))
@@ -106,10 +111,10 @@ func Test_Case5(t *testing.T) {
 	seed := make([]byte, 8)
 	binary.LittleEndian.PutUint64(seed, 500)
 
-	candidates, flipsPerAuthor, flips := makeCandidatesWithFlips(seed, 10, 3, 3)
+	candidates := makeCandidatesWithFlips(seed, 10, 3, 3)
 
-	authorsPerCandidate, _ := GetAuthorsDistribution(candidates, seed, 7)
-	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates), authorsPerCandidate, flipsPerAuthor, flips, seed, 7)
+	authorsPerCandidate := GetAuthorsDistribution(candidates, seed, 7)
+	shortFlipsPerCandidate, longFlipsPerCandidate := GetFlipsDistribution(len(candidates[0].candidates), authorsPerCandidate[0].authorsPerCandidate, candidates[0].flipsPerAuthor, candidates[0].flips, seed, 7)
 
 	for _, item := range shortFlipsPerCandidate {
 		require.True(t, len(item) >= 6)
@@ -139,17 +144,14 @@ func Test_FillAuthorsQueue(t *testing.T) {
 
 func makeCandidates(authors int) []*candidate {
 	res := make([]*candidate, 0)
-	b := random.GetRandomBytes(12)
 	for i := 0; i < authors; i++ {
 		res = append(res, &candidate{
-			Code:       b,
-			Generation: 0,
 		})
 	}
 	return res
 }
 
-func makeCandidatesWithFlips(seed []byte, candidatesCount int, authorsCount int, flipPerAuthor int) (candidates []*candidate, flipsPerAuthor map[int][][]byte, flips [][]byte) {
+func makeCandidatesWithFlips(seed []byte, candidatesCount int, authorsCount int, flipPerAuthor int) (candidatesByShard map[common.ShardId]*candidatesOfShard) {
 	res := make([]*candidate, 0)
 	randSeed := binary.LittleEndian.Uint64(seed)
 	//code := random.GetRandomBytes(12)
@@ -160,8 +162,8 @@ func makeCandidatesWithFlips(seed []byte, candidatesCount int, authorsCount int,
 		authorsMap[a] = true
 	}
 
-	flipsPerAuthor = make(map[int][][]byte)
-	flips = make([][]byte, 0)
+	flipsPerAuthor := make(map[int][][]byte)
+	flips := make([][]byte, 0)
 
 	for i := 0; i < candidatesCount; i++ {
 		isAuthor := authorsMap[i]
@@ -173,10 +175,16 @@ func makeCandidatesWithFlips(seed []byte, candidatesCount int, authorsCount int,
 			}
 		}
 		res = append(res, &candidate{
-			Code:       random.GetRandomBytes(12),
-			Generation: 0,
-			IsAuthor:   isAuthor,
+			IsAuthor: isAuthor,
 		})
 	}
-	return res, flipsPerAuthor, flips
+
+	candidatesByShard = make(map[common.ShardId]*candidatesOfShard)
+	candidatesByShard[0] = &candidatesOfShard{
+		candidates:     res,
+		flipsPerAuthor: flipsPerAuthor,
+		flips:          flips,
+	}
+
+	return candidatesByShard
 }