+++ events/events.go
@@ -2,9 +2,11 @@ package events
 
 import (
 	"github.com/idena-network/idena-go/blockchain/types"
+	"github.com/idena-network/idena-go/common"
 	"github.com/idena-network/idena-go/common/eventbus"
 	iface "github.com/ipfs/interface-go-ipfs-core"
 	"github.com/libp2p/go-libp2p-core"
+	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"time"
 )
 
@@ -18,11 +20,13 @@ const (
 	IpfsPortChangedEventId = eventbus.EventID("ipfs-port-changed")
 	DeleteFlipEventID      = eventbus.EventID("flip-delete")
 	PeersEventID           = eventbus.EventID("peers")
+	BlockchainResetEventID = eventbus.EventID("chain-reset")
 )
 
 type NewTxEvent struct {
 	Tx       *types.Transaction
 	Own      bool
+	ShardId  common.ShardId
 	Deferred bool
 }
 
@@ -39,8 +43,9 @@ func (e *NewBlockEvent) EventID() eventbus.EventID {
 }
 
 type NewFlipKeyEvent struct {
-	Key *types.PublicFlipKey
-	Own bool
+	Key     *types.PublicFlipKey
+	ShardId common.ShardId
+	Own     bool
 }
 
 func (e *NewFlipKeyEvent) EventID() eventbus.EventID {
@@ -63,8 +68,9 @@ func (NewFlipEvent) EventID() eventbus.EventID {
 }
 
 type NewFlipKeysPackageEvent struct {
-	Key *types.PrivateFlipKeysPackage
-	Own bool
+	Key     *types.PrivateFlipKeysPackage
+	ShardId common.ShardId
+	Own     bool
 }
 
 func (e *NewFlipKeysPackageEvent) EventID() eventbus.EventID {
@@ -72,7 +78,8 @@ func (e *NewFlipKeysPackageEvent) EventID() eventbus.EventID {
 }
 
 type IpfsPortChangedEvent struct {
-	Host core.Host
+	Host   core.Host
+	PubSub *pubsub.PubSub
 }
 
 func (i IpfsPortChangedEvent) EventID() eventbus.EventID {
@@ -94,4 +101,13 @@ type PeersEvent struct {
 
 func (e *PeersEvent) EventID() eventbus.EventID {
 	return PeersEventID
+}
+
+type BlockchainResetEvent struct {
+	Header      *types.Header
+	RevertedTxs []*types.Transaction
+}
+
+func (e *BlockchainResetEvent) EventID() eventbus.EventID {
+	return BlockchainResetEventID
 }