+++ node/node.go
@@ -186,7 +186,7 @@ func NewNodeWithInjections(config *config.Config, bus eventbus.Bus, statsCollect
 	chain := blockchain.NewBlockchain(config, db, txpool, appState, ipfsProxy, secStore, bus, offlineDetector, keyStore, subManager, upgrader)
 	proposals, pendingProofs := pengings.NewProposals(chain, appState, offlineDetector, upgrader)
 	flipper := flip.NewFlipper(db, ipfsProxy, flipKeyPool, txpool, secStore, appState, bus)
-	pm := protocol.NewIdenaGossipHandler(ipfsProxy.Host(), config.P2P, chain, proposals, votes, txpool, flipper, bus, flipKeyPool, appVersion, &ceremonyChecker{
+	pm := protocol.NewIdenaGossipHandler(ipfsProxy.Host(), ipfsProxy.PubSub(), config.P2P, chain, proposals, votes, txpool, flipper, bus, flipKeyPool, appVersion, &ceremonyChecker{
 		appState: appState,
 		chain:    chain,
 	})
@@ -275,7 +275,7 @@ func (node *Node) StartWithHeight(height uint64) {
 	}
 
 	if height > 0 && node.blockchain.Head.Height() > height {
-		if err := node.blockchain.ResetTo(height); err != nil {
+		if _, err := node.blockchain.ResetTo(height); err != nil {
 			node.log.Error(fmt.Sprintf("Cannot reset blockchain to %d", height), "error", err.Error())
 			return
 		}